backend/requirements.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# LangChain & LangGraph
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
langchain-community==0.0.10

# Google AI
google-generativeai==0.3.2
google-cloud-vision==3.5.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0

# Vector Search & Retrieval
numpy==1.26.2
scikit-learn==1.3.2
rank-bm25==0.2.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
pre-commit==3.6.0

# Utilities
python-magic==0.4.27
tqdm==4.66.1
